etl/requirements.txt
elastic-transport==8.4.0
elasticsearch==8.7.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0